Fixing race condition with ref setting
